training/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "artemis-training"
version = "0.1.0"
description = "Training pipeline for Artemis code generation model (Qwen2.5-Coder-7B QLoRA)"
requires-python = ">=3.10"
dependencies = [
    # Data collection
    "google-cloud-bigquery>=3.14",
    "datasets>=2.16",
    "huggingface-hub>=0.20",
    # Preprocessing
    "tree-sitter>=0.21",
    "tree-sitter-go>=0.21",
    "datasketch>=1.6",
    "pandas>=2.1",
    # Training
    "torch>=2.1",
    "transformers>=4.38",
    "peft>=0.8",
    "trl>=0.7",
    "bitsandbytes>=0.42",
    "accelerate>=0.26",
    # Evaluation
    "rouge-score>=0.1",
    "sacrebleu>=2.4",
    # Deployment
    "autoawq>=0.2",
    # Utilities
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "wandb>=0.16",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = ["ruff>=0.2", "pytest>=8.0"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
